refactor: move 패키지 구성
