+++ src/main/java/racingcar/model/move/Move.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.model.move;
 
 public interface Move {
     boolean isMove(int number);
